mem fix

+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: mnanke <mnanke@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/24 23:20:00 by mnanke            #+#    #+#             */
-/*   Updated: 2023/01/28 18:07:40 by mnanke           ###   ########.fr       */
+/*   Updated: 2023/02/14 15:33:24 by mnanke           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t	i;
+	long long int	i;
 
 	i = ft_strlen(s);
 	while (0 <= i)
